Setup tailwind purge for build, add anchor links to each app
