+++ apps/admin/src/app/admin-login/admin-login.component.html
@@ -1,3 +1,8 @@
 <div class="max-w-screen-sm flex h-full mx-auto">
-    <shared-login class="m-auto" [title]="loginTitle"></shared-login>
+    <div class="m-auto">
+        <shared-login [title]="loginTitle"></shared-login>
+        <p class="mt-4">
+            <a class="text-blue-400 hover:underline" href="./">Client Login</a>
+        </p>
+    </div>
 </div>
