+++ apps/client/src/app/client-login/client-login.component.html
@@ -1,3 +1,9 @@
 <div class="max-w-screen-sm flex h-full mx-auto">
-    <shared-login class="m-auto" [title]="loginTitle"></shared-login>
+    <div class="m-auto">
+        <shared-login class="m-auto" [title]="loginTitle"></shared-login>
+        <p class="mt-4">
+            <a class="text-blue-400 hover:underline" href="./admin">Admin Login</a>
+        </p>
+    </div>
+    
 </div>
