+++ apps/client/src/styles.scss
@@ -1,4 +1,1 @@
 @import '~@angular/material/prebuilt-themes/deeppurple-amber.css';
-@import 'tailwindcss/base';
-@import 'tailwindcss/components';
-@import 'tailwindcss/utilities';
