+++ package.json
@@ -51,6 +51,7 @@
     "@angular/cli": "~10.1.3",
     "@angular/compiler-cli": "^10.1.0",
     "@angular/language-service": "^10.1.0",
+    "@fullhuman/postcss-purgecss": "^3.0.0",
     "@nrwl/cli": "10.4.4",
     "@nrwl/cypress": "10.4.4",
     "@nrwl/jest": "10.4.4",
