+++ tailwind.config.js
@@ -1,5 +1,20 @@
 module.exports = {
-  purge: [],
+  purge:{
+    content:[
+      "./apps/**/*.html",
+      "./apps/*/*.ts",
+      "./apps/**/*.scss",
+      "./libs/**/*.html",
+      "./libs/**/*.ts"],
+    options:{
+      rejected: true,
+      printRejected: true,
+      whitelistPatterns: [
+        /^cdk-/, 
+        /^mat-/, 
+      ],
+    }
+  },
   darkMode: false, // or 'media' or 'class'
   theme: {
     extend: {},
