+++ webpack.config.js
@@ -1,11 +1,23 @@
 "use strict";
 
-const webpackCommonConfig = require("./webpack-common.config");
+const helpers = require("./helpers");
+const tailwindWebpackRule = {
+  test: /tailwind\.scss$/,
+  loader: "postcss-loader",
+  options: {
+    ident: "postcss",
+    syntax: "postcss-scss",
+    plugins: () => [
+      require("tailwindcss")(helpers.root("tailwind.config.js")), 
+      require('autoprefixer'),
+    ],
+  },
+};
 
 module.exports = {
   module: {
     rules: [
-      webpackCommonConfig.tailwindWebpackRule,
+      tailwindWebpackRule,
     ],
   },
 };
